Enhance course and exam management: add caching updates and improve error handling
