+++ admin_service/controller/dbController.js
@@ -1,5 +1,6 @@
 import pool from "../db/connectDb.js";
-import { insertAllCourseCache } from "../redis/queries.js";
+import { selectCourseExamsQuery } from "../db/queries.js";
+import { insertAllCourseCache, setCourseExamsCache } from "../redis/queries.js";
 import { ResponseBuilder } from "../utils/response.js";
 
 export const getCoursesFromDB = async (res) => {
@@ -8,6 +9,23 @@ export const getCoursesFromDB = async (res) => {
     const [rows] = await pool.query(query);
     if (rows.length) {
       insertAllCourseCache(rows);
+      ResponseBuilder.successResponse(res, rows);
+    } else {
+      ResponseBuilder.errorResponse(res);
+    }
+  } catch (error) {
+    error.sql
+      ? ResponseBuilder.sqlerrorResponse(res, { ...error })
+      : ResponseBuilder.errorResponse(res);
+  }
+};
+
+export const getCourseExamsDB = async (res, course_id, semester) => {
+  try {
+    const query = selectCourseExamsQuery();
+    const [rows] = await pool.query(query, [course_id, semester]);
+    if (rows.length) {
+      setCourseExamsCache(rows);
       ResponseBuilder.successResponse(res, rows);
     } else {
       ResponseBuilder.errorResponse(res);
